+++ app/src/main/java/com/r0adkll/deckbuilder/arch/data/DataModule.kt
@@ -12,7 +12,9 @@ import com.r0adkll.deckbuilder.arch.data.features.account.DefaultAccountReposito
 import com.r0adkll.deckbuilder.arch.data.features.cards.cache.CardCache
 import com.r0adkll.deckbuilder.arch.data.features.cards.cache.RoomCardCache
 import com.r0adkll.deckbuilder.arch.data.features.cards.repository.DefaultCardRepository
+import com.r0adkll.deckbuilder.arch.data.features.cards.repository.source.CachingNetworkSearchDataSource
 import com.r0adkll.deckbuilder.arch.data.features.cards.repository.source.CombinedSearchDataSource
+import com.r0adkll.deckbuilder.arch.data.features.cards.repository.source.DiskSearchDataSource
 import com.r0adkll.deckbuilder.arch.data.features.cards.repository.source.SearchDataSource
 import com.r0adkll.deckbuilder.arch.data.features.community.cache.CommunityCache
 import com.r0adkll.deckbuilder.arch.data.features.community.cache.FirestoreCommunityCache
@@ -39,6 +41,7 @@ import com.r0adkll.deckbuilder.arch.data.features.validation.model.DuplicateRule
 import com.r0adkll.deckbuilder.arch.data.features.validation.model.PrismStarRule
 import com.r0adkll.deckbuilder.arch.data.features.validation.model.SizeRule
 import com.r0adkll.deckbuilder.arch.data.features.validation.repository.DefaultDeckValidator
+import com.r0adkll.deckbuilder.arch.data.remote.FirebaseRemote
 import com.r0adkll.deckbuilder.arch.data.remote.plugin.CacheInvalidatePlugin
 import com.r0adkll.deckbuilder.arch.data.remote.plugin.RemotePlugin
 import com.r0adkll.deckbuilder.arch.domain.features.account.AccountRepository
@@ -52,11 +55,13 @@ import com.r0adkll.deckbuilder.arch.domain.features.importer.repository.Importer
 import com.r0adkll.deckbuilder.arch.domain.features.missingcard.repository.MissingCardRepository
 import com.r0adkll.deckbuilder.arch.domain.features.offline.repository.OfflineRepository
 import com.r0adkll.deckbuilder.arch.domain.features.preview.PreviewRepository
+import com.r0adkll.deckbuilder.arch.domain.features.remote.Remote
 import com.r0adkll.deckbuilder.arch.domain.features.testing.DeckTester
 import com.r0adkll.deckbuilder.arch.domain.features.validation.model.Rule
 import com.r0adkll.deckbuilder.arch.domain.features.validation.repository.DeckValidator
 import com.r0adkll.deckbuilder.internal.di.scopes.AppScope
 import com.r0adkll.deckbuilder.util.Schedulers
+import com.r0adkll.deckbuilder.util.helper.Connectivity
 import dagger.Module
 import dagger.Provides
 import dagger.multibindings.ElementsIntoSet
@@ -86,6 +91,10 @@ class DataModule {
     fun provideRxSharedPreferences(sharedPreferences: SharedPreferences) : RxSharedPreferences {
         return RxSharedPreferences.create(sharedPreferences)
     }
+
+
+    @Provides @AppScope
+    fun provideRemotePreferences(remoteConfig: FirebaseRemote): Remote = remoteConfig
 
 
     @Provides @AppScope
@@ -155,7 +164,19 @@ class DataModule {
 
 
     @Provides @AppScope
-    fun provideSearchDataSource(dataSource: CombinedSearchDataSource): SearchDataSource = dataSource
+    fun provideSearchDataSource(
+            api: Pokemon,
+            cache: CardCache,
+            source: ExpansionDataSource,
+            remote: Remote,
+            schedulers: Schedulers,
+            preferences: AppPreferences,
+            connectivity: Connectivity
+    ): SearchDataSource {
+        val disk: SearchDataSource = DiskSearchDataSource(cache, schedulers)
+        val network: SearchDataSource = CachingNetworkSearchDataSource(api, source, cache, remote, schedulers)
+        return CombinedSearchDataSource(preferences, disk, network, connectivity)
+    }
 
 
     /*