+++ app/src/main/java/com/r0adkll/deckbuilder/arch/data/features/cards/repository/source/CombinedSearchDataSource.kt
@@ -1,7 +1,7 @@
 package com.r0adkll.deckbuilder.arch.data.features.cards.repository.source
 
 import com.r0adkll.deckbuilder.arch.data.AppPreferences
-import com.r0adkll.deckbuilder.arch.data.remote.Remote
+import com.r0adkll.deckbuilder.arch.domain.features.remote.Remote
 import com.r0adkll.deckbuilder.arch.data.features.cards.cache.CardCache
 import com.r0adkll.deckbuilder.arch.data.features.expansions.ExpansionDataSource
 import com.r0adkll.deckbuilder.arch.domain.features.cards.model.Filter
@@ -11,48 +11,46 @@ import com.r0adkll.deckbuilder.util.helper.Connectivity
 import io.pokemontcg.Pokemon
 import io.pokemontcg.model.SuperType
 import io.reactivex.Observable
+import io.reactivex.functions.Function
 import javax.inject.Inject
 
 
-class CombinedSearchDataSource @Inject constructor(
-        val api: Pokemon,
-        val cache: CardCache,
-        val source: ExpansionDataSource,
-        val remote: Remote,
+class CombinedSearchDataSource(
         val preferences: AppPreferences,
-        val schedulers: Schedulers,
+        val diskSource: SearchDataSource,
+        val networkSource: SearchDataSource,
         val connectivity: Connectivity
 ) : SearchDataSource {
 
-    private val disk: SearchDataSource = DiskSearchDataSource(cache, schedulers)
-    private val network: SearchDataSource = CachingNetworkSearchDataSource(api, source, cache, remote, schedulers)
-
-
     override fun search(type: SuperType?, query: String, filter: Filter?): Observable<List<PokemonCard>> {
-        val forceDiskSearch = filter?.expansions?.map { it.code }?.let { preferences.offlineExpansions.get().containsAll(it) } ?: false
+        val forceDiskSearch = filter?.expansions
+                ?.map { it.code }
+                ?.let { it.isNotEmpty() && preferences.offlineExpansions.get().containsAll(it) }
+                ?: false
         return if (connectivity.isConnected() && !forceDiskSearch) {
-            network.search(type, query, filter)
-                    .onErrorResumeNext(disk.search(type, query, filter))
+            networkSource.search(type, query, filter)
+                    .onErrorResumeNext(Function {
+                        diskSource.search(type, query, filter)
+                    })
         } else {
-            disk.search(type, query, filter)
+            diskSource.search(type, query, filter)
         }
     }
 
-
     override fun find(ids: List<String>): Observable<List<PokemonCard>> {
         return if (connectivity.isConnected()) {
-            disk.find(ids)
+            diskSource.find(ids)
                     .flatMap { diskCards ->
                         val missingIds = ids.filter { id -> diskCards.none { card -> card.id == id } }
                         if (missingIds.isNotEmpty()) {
-                            network.find(missingIds)
+                            networkSource.find(missingIds)
                                     .map { it.plus(diskCards) }
                         } else {
                             Observable.just(diskCards)
                         }
                     }
         } else {
-            disk.find(ids)
+            diskSource.find(ids)
         }
     }
 }