+++ app/src/main/java/com/r0adkll/deckbuilder/arch/domain/features/decks/model/Deck.kt
@@ -1,14 +1,14 @@
 package com.r0adkll.deckbuilder.arch.domain.features.decks.model
 
 
+import android.os.Parcelable
 import com.r0adkll.deckbuilder.arch.domain.features.cards.model.PokemonCard
 import com.r0adkll.deckbuilder.arch.ui.features.deckbuilder.deckimage.adapter.DeckImage
 import io.pokemontcg.model.SuperType
-import paperparcel.PaperParcel
-import paperparcel.PaperParcelable
+import kotlinx.android.parcel.Parcelize
 
 
-@PaperParcel
+@Parcelize
 data class Deck(
         val id: String,
         val name: String,
@@ -17,7 +17,7 @@ data class Deck(
         val cards: List<PokemonCard>,
         val isMissingCards: Boolean,
         val timestamp: Long
-) : PaperParcelable {
+) : Parcelable {
 
     val pokemonCount: Int get() = cards.count { it.supertype == SuperType.POKEMON }
     val trainerCount: Int get() = cards.count { it.supertype == SuperType.TRAINER }
@@ -25,9 +25,5 @@ data class Deck(
 
     override fun toString(): String {
         return "Deck(id='$id', name='$name', description='$description', cards=${cards.size}, isMissingCards=$isMissingCards, timestamp=$timestamp)"
-    }
-
-    companion object {
-        @JvmField val CREATOR = PaperParcelDeck.CREATOR
     }
 }