+++ app/src/main/java/com/r0adkll/deckbuilder/arch/domain/features/editing/model/Session.kt
@@ -1,13 +1,13 @@
 package com.r0adkll.deckbuilder.arch.domain.features.editing.model
 
 
+import android.os.Parcelable
 import com.r0adkll.deckbuilder.arch.domain.features.cards.model.PokemonCard
 import com.r0adkll.deckbuilder.arch.ui.features.deckbuilder.deckimage.adapter.DeckImage
-import paperparcel.PaperParcel
-import paperparcel.PaperParcelable
+import kotlinx.android.parcel.Parcelize
 
 
-@PaperParcel
+@Parcelize
 data class Session(
         val id: Long,
         val deckId: String?,
@@ -17,17 +17,14 @@ data class Session(
         val cards: List<PokemonCard>,
         val hasChanges: Boolean,
         val changes: List<Change>
-) : PaperParcelable {
-
+) : Parcelable {
 
     override fun toString(): String {
         return "Session(id=$id, deckId=$deckId, name='$name', description='$description', " +
                 "cards=${cards.size}, hasChanges=$hasChanges, changes=${changes.size})"
     }
 
-
     companion object {
-        @JvmField val CREATOR = PaperParcelSession.CREATOR
 
         const val NO_ID = -1L
     }