+++ app/src/main/java/com/r0adkll/deckbuilder/arch/ui/components/preview/ExpansionPreviewRenderer.kt
@@ -11,8 +11,9 @@ import android.util.Base64
 import android.view.View
 import android.widget.ImageView
 import com.r0adkll.deckbuilder.GlideApp
-import com.r0adkll.deckbuilder.arch.data.remote.model.ExpansionPreview.PreviewSpec
+import com.r0adkll.deckbuilder.arch.domain.features.remote.model.ExpansionPreview.PreviewSpec
 import com.r0adkll.deckbuilder.arch.ui.widgets.AspectRatioImageView
+import com.r0adkll.deckbuilder.arch.ui.widgets.BackgroundDrawableWrapper
 import com.r0adkll.deckbuilder.util.extensions.margins
 import com.r0adkll.deckbuilder.util.glide.AlphaTransformation
 import com.r0adkll.deckbuilder.util.glide.RxGlide.asObservable
@@ -48,7 +49,7 @@ object ExpansionPreviewRenderer {
     }
 
 
-    fun applyForeground(imageView: ImageView, spec: PreviewSpec.DrawableSpec): Disposable {
+    fun applyForeground(imageView: ImageView, spec: PreviewSpec.DrawableSpec): Disposable? {
         // Apply options
         imageView.margins(spec.margins)
         spec.alpha?.let { imageView.alpha = it }
@@ -62,14 +63,27 @@ object ExpansionPreviewRenderer {
             }
         }
 
-        // Load Drawable
-        return createDrawable(imageView.context, spec)
-                .observeOn(AndroidSchedulers.mainThread())
-                .subscribe({
-                    imageView.setImageDrawable(it)
-                }, {
-                    Timber.e(it, "Error loading foreground drawable")
-                })
+        // Due to Issue #56 - If it's an URL shortcut to using glide
+        if (spec.source.type == "url") {
+            var request = GlideApp.with(imageView)
+                    .load(spec.source.value)
+
+            spec.alpha?.let {
+                request = request.transform(AlphaTransformation(it))
+            }
+
+            request.into(imageView)
+
+            return null
+        } else {
+            return createDrawable(imageView.context, spec)
+                    .observeOn(AndroidSchedulers.mainThread())
+                    .subscribe({
+                        imageView.setImageDrawable(it)
+                    }, {
+                        Timber.e(it, "Error loading foreground drawable")
+                    })
+        }
     }
 
 
@@ -102,6 +116,7 @@ object ExpansionPreviewRenderer {
                         }
                     } as Drawable
                 }
+                .map { BackgroundDrawableWrapper(it) as Drawable }
                 .subscribeOn(Schedulers.computation())
     }