+++ app/src/main/java/com/r0adkll/deckbuilder/arch/ui/features/decks/adapter/Item.kt
@@ -2,7 +2,7 @@ package com.r0adkll.deckbuilder.arch.ui.features.decks.adapter
 
 
 import com.r0adkll.deckbuilder.R
-import com.r0adkll.deckbuilder.arch.data.remote.model.ExpansionPreview
+import com.r0adkll.deckbuilder.arch.domain.features.remote.model.ExpansionPreview
 import com.r0adkll.deckbuilder.arch.domain.features.decks.model.Deck
 import com.r0adkll.deckbuilder.arch.ui.components.RecyclerItem
 import com.r0adkll.deckbuilder.arch.ui.features.decks.DecksUi