+++ app/src/main/java/com/r0adkll/deckbuilder/arch/ui/features/settings/SettingsActivity.kt
@@ -27,6 +27,7 @@ import com.r0adkll.deckbuilder.arch.domain.features.account.AccountRepository
 import com.r0adkll.deckbuilder.arch.ui.Shortcuts
 import com.r0adkll.deckbuilder.arch.ui.components.BaseActivity
 import com.r0adkll.deckbuilder.arch.ui.components.BasePreferenceFragment
+import com.r0adkll.deckbuilder.arch.ui.features.settings.offline.ManageOfflineActivity
 import com.r0adkll.deckbuilder.arch.ui.features.setup.SetupActivity
 import com.r0adkll.deckbuilder.internal.analytics.Analytics
 import com.r0adkll.deckbuilder.internal.analytics.Event
@@ -151,6 +152,7 @@ class SettingsActivity : BaseActivity() {
                 }
                 "pref_cache_manage" -> {
                     Analytics.event(Event.SelectContent.Action("settings", "manage_cache"))
+                    startActivity(ManageOfflineActivity.createIntent(activity!!))
                     true
                 }
                 "pref_account_signout" -> {
@@ -296,16 +298,23 @@ class SettingsActivity : BaseActivity() {
                                     setupPreferences()
 
                                     // Now we need to migrate any existing local decks to their account
-                                    migrationSnackbar = Snackbar.make(view!!, R.string.account_migration_started, Snackbar.LENGTH_INDEFINITE)
-                                    migrationSnackbar?.show()
+                                    view?.let {
+                                        migrationSnackbar = Snackbar.make(it, R.string.account_migration_started, Snackbar.LENGTH_INDEFINITE)
+                                        migrationSnackbar?.show()
+                                    }
+
                                     disposables += accountRepository.migrateAccount()
                                             .observeOn(AndroidSchedulers.mainThread())
                                             .subscribe({
-                                                migrationSnackbar = Snackbar.make(view!!, R.string.account_migration_finished, Snackbar.LENGTH_SHORT)
-                                                migrationSnackbar?.show()
-                                            }, {
-                                                migrationSnackbar = Snackbar.make(view!!, it.localizedMessage, Snackbar.LENGTH_SHORT)
-                                                migrationSnackbar?.show()
+                                                view?.let { v ->
+                                                    migrationSnackbar = Snackbar.make(v, R.string.account_migration_finished, Snackbar.LENGTH_SHORT)
+                                                    migrationSnackbar?.show()
+                                                }
+                                            }, { e ->
+                                                view?.let { v ->
+                                                    migrationSnackbar = Snackbar.make(v, e.localizedMessage, Snackbar.LENGTH_SHORT)
+                                                    migrationSnackbar?.show()
+                                                }
                                             })