+++ app/src/main/java/com/r0adkll/deckbuilder/arch/ui/features/settings/offline/ManageOfflineActivity.kt
@@ -1,5 +1,7 @@
 package com.r0adkll.deckbuilder.arch.ui.features.settings.offline
 
+import android.content.Context
+import android.content.Intent
 import android.os.Bundle
 import com.r0adkll.deckbuilder.R
 import com.r0adkll.deckbuilder.arch.ui.components.BaseActivity
@@ -11,11 +13,21 @@ class ManageOfflineActivity : BaseActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_manage_offline)
+
+        supportActionBar?.setDisplayHomeAsUpEnabled(true)
+        appbar?.setNavigationOnClickListener { finish() }
     }
 
 
     override fun setupComponent(component: AppComponent) {
-        TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
+
     }
 
+
+    companion object {
+
+        fun createIntent(context: Context): Intent {
+            return Intent(context, ManageOfflineActivity::class.java)
+        }
+    }
 }